+++ NEWS.md
@@ -1,4 +1,6 @@
 # aion 1.0.0.9000
+## Bugfixes & changes
+* `plot()` now supports missing values.
 
 # aion 1.0.0